Moved fonts, svgs and updated download script
