+++ components/features/DesignPanel.vue
@@ -6,8 +6,8 @@ import type { Canvas } from 'fabric'
 import { ref } from 'vue'
 import { useCanvasStore } from '~/stores/canvasStore'
 import CanvasPanel from '~/components/features/CanvasPanel.vue'
-import ImageIcon from '~/assets/images/custom-design/image-icon.svg?component'
-import TextIcon from '~/assets/images/custom-design/text-icon.svg?component'
+import ImageIcon from '~/assets/images/custom-design/textile-advertisement-decor/image-icon.svg?component'
+import TextIcon from '~/assets/images/custom-design/textile-advertisement-decor/text-icon.svg?component'
 import BackgroundSelector from '~/components/features/BackgroundSelector.vue'
 import TextboxControls from '~/components/features/TextboxControls.vue'
 import IconButton from '~/components/common/IconButton.vue'
